+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Fixed
+- proxy: handle empty ct parameter
 
 ## [0.0.2] - 2018-01-16
 ### Added